- name fix android and simplify types on ios
